Move DFS to other file
