--- packages/react-native-gesture-handler/src/v3/HostGestureDetector.tsx
@@ -1,3 +1,0 @@
-import RNGestureHandlerDetectorNativeComponent from '../specs/RNGestureHandlerDetectorNativeComponent';
-const HostGestureDetector = RNGestureHandlerDetectorNativeComponent;
-export default HostGestureDetector;
